+++ typescript/infra/scripts/sealevel-helpers/deploy-svm-ism.ts
@@ -54,7 +54,7 @@ import { Contexts } from '../../config/contexts.js';
 import { getChain } from '../../config/registry.js';
 import { getDeployerKey } from '../../src/agents/key-utils.js';
 import { chainsToSkip } from '../../src/config/chain.js';
-import { DeployEnvironment } from '../../src/config/environment.js';
+import { DeployEnvironment } from '../../src/config/deploy-environment.js';
 import { squadsConfigs } from '../../src/config/squads.js';
 import {
   SvmMultisigConfigMap,