+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @hyperlane-xyz/helloworld
 
+## 33.0.2
+
+### Patch Changes
+
+- Updated dependencies [1f918d0]
+- Updated dependencies [78199f4]
+  - @hyperlane-xyz/sdk@33.0.2
+  - @hyperlane-xyz/utils@33.0.2
+  - @hyperlane-xyz/core@11.3.1
+
 ## 33.0.1
 
 ### Patch Changes