+++ typescript/keyfunder/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @hyperlane-xyz/keyfunder
 
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [1f918d0]
+- Updated dependencies [78199f4]
+  - @hyperlane-xyz/sdk@33.0.2
+  - @hyperlane-xyz/utils@33.0.2
+  - @hyperlane-xyz/metrics@0.2.20
+  - @hyperlane-xyz/core@11.3.1
+
 ## 0.1.26
 
 ### Patch Changes